Bug #1409373 trace: coverity waring overflow
[root cause  ] unsigned long is mismatch with u64
[changes     ] force the var into u64
[side effects] no
[self test   ]:yes
[download normally]:yes
[power on/off normally]:yes
[do common repository/branch inspection]:yes
[is there dependence]:no
[confirm dependent commit]:yes
[board]:SC9832E_ANDROID11_TRUNK
[test case]:no
[reviewers   ]
[change_type ] feature_bugfix
[tag_product ] common

the var latency's type is unsigned long, but it was assiged to an u64
var, it may be overflow. so force the var into u64 while calculating

+++ kernel/trace/trace_irqsoff_bytedancy.c
@@ -455,7 +455,7 @@ static ssize_t trace_latency_write(struct file *file, const char __user *buf,
 	} else if (latency < sampling_period_ms)
 		return -EINVAL;
 
-	trace_irqoff_latency = latency * 1000 * 1000UL;
+	trace_irqoff_latency = (u64)latency * 1000 * 1000UL;
 
 	return count;
 }
